static/bill/crm420update-main/crm420update-main/requirements.txt
Flask
Flask-SQLAlchemy
Flask-Login
Flask-JWT-Extended
Flask-WTF
Werkzeug
pandas
openpyxl
Pillow
python-barcode
qrcode
WTForms
email-validator
gunicorn
